Homeworks 21,22
