+++ api-tests/src/main/java/org/example/cartData/models/AddCartItemRequest.java
@@ -1,0 +1,13 @@
+package org.example.cartData.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class AddCartItemRequest {
+
+    private String productId;
+    private int countOfProducts;
+
+}
