+++ api-tests/src/main/java/org/example/cartData/models/CartItem.java
@@ -1,0 +1,19 @@
+package org.example.cartData.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CartItem {
+
+    private String id;
+    private int quantity;
+    private Double discountPercentage;
+    private String cartId;
+    private String productId;
+    private Product product;
+
+}
