+++ api-tests/src/main/java/org/example/cartData/models/CreateCartResponse.java
@@ -1,0 +1,9 @@
+package org.example.cartData.models;
+
+import lombok.Data;
+
+@Data
+public class CreateCartResponse {
+
+    String id;
+}
