+++ api-tests/src/main/java/org/example/cartData/models/Product.java
@@ -1,0 +1,21 @@
+package org.example.cartData.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Product {
+
+    private String id;
+    private String name;
+    private String description;
+    private double price;
+    private boolean isLocationOffer;
+    private boolean isRental;
+    private boolean inStock;
+}
