+++ api-tests/src/main/java/org/example/cartData/models/UpdateCartResponse.java
@@ -1,0 +1,10 @@
+package org.example.cartData.models;
+
+import lombok.Data;
+
+@Data
+public class UpdateCartResponse {
+
+    String message;
+
+}
