+++ api-tests/src/main/java/org/example/userData/models/ChangePasswordRequest.java
@@ -1,0 +1,12 @@
+package org.example.userData.models;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ChangePasswordRequest {
+    private String currentPassword;
+    private String newPassword;
+    private String newPasswordConfirmation;
+}
