+++ api-tests/src/main/java/org/example/userData/models/ChangePasswordResponse.java
@@ -1,0 +1,10 @@
+package org.example.userData.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class ChangePasswordResponse {
+    private String message;
+}
