+++ api-tests/src/main/java/org/example/userData/models/LoginRequest.java
@@ -1,0 +1,12 @@
+package org.example.userData.models;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+
+@Data
+@AllArgsConstructor
+public class LoginRequest {
+    private String password;
+    private String email;
+}
