+++ api-tests/src/main/java/org/example/userData/models/LoginResponse.java
@@ -1,0 +1,12 @@
+package org.example.userData.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class LoginResponse {
+
+        private String accessToken;
+        private String tokenType;
+        private int expiresIn;
+}
