--- src/main/java/org/example/cartData/models/AddCartItemRequest.java
@@ -1,12 +1,0 @@
-package org.example.cartData.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class AddCartItemRequest {
-
-    private String productId;
-    private int countOfProducts;
-}
