--- src/main/java/org/example/cartData/models/CreateCartResponse.java
@@ -1,10 +1,0 @@
-package org.example.cartData.models;
-
-import lombok.Data;
-
-@Data
-public class CreateCartResponse {
-
-    String id;
-
-}
