--- src/main/java/org/example/cartData/models/UpdateCartResponse.java
@@ -1,9 +1,0 @@
-package org.example.cartData.models;
-
-import lombok.Data;
-
-@Data
-public class UpdateCartResponse {
-
-    String message;
-}
