--- src/main/java/org/example/userData/models/RegisterUserResponse.java
@@ -1,24 +1,0 @@
-package org.example.userData.models;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class RegisterUserResponse {
-
-    private String firstName;
-    private String lastName;
-    private String address;
-    private String city;
-    private String state;
-    private String country;
-    private String postcode;
-    private String phone;
-    private String dob;
-    private String email;
-    private String id;
-    private String createdAt;
-
-}
